rustlings solution

+++ exercises/conversions/using_as.rs
@@ -1,16 +1,13 @@
-// Type casting in Rust is done via the usage of the `as` operator.
-// Please note that the `as` operator is not only used when type casting.
-// It also helps with renaming imports.
+// Rust 中的类型转换是通过 `as` 运算符完成的。
+// 请注意 `as` 运算符不止用于类型转换，
+// 它还有助于重命名导入。
 //
-// The goal is to make sure that the division does not fail to compile
-// and returns the proper type.
+// 目标是确保除法不会编译失败，并返回合适的类型。
 // 执行 `rustlings hint using_as` 或在观察模式下使用 `hint` 子命令来获取提示。
-
-// I AM NOT DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64
 }
 
 fn main() {
